fix: move prize view move to its package
